+++ src/js/components/slideshow.js
@@ -57,12 +57,13 @@
 
             var $this = this, canvas;
 
-            this.container   = UI.$(this.find('.@-slideshow').andSelf().filter(UI.prefix('.@-slideshow'))),
-            this.slides      = this.container.children(),
-            this.slidesCount = this.slides.length,
-            this.current     = this.options.start,
-            this.animating   = false,
-            this.triggers    = this.find('[data-@-slideshow-item]');
+            this.container     = UI.$(this.find('.@-slideshow').andSelf().filter(UI.prefix('.@-slideshow')));
+            this.slides        = this.container.children();
+            this.slidesCount   = this.slides.length;
+            this.current       = this.options.start;
+            this.animating     = false;
+            this.triggers      = this.find('[data-@-slideshow-item]');
+            this.fixFullscreen = navigator.userAgent.match(/(iPad|iPhone|iPod)/g) && this.container.hasClass(UI.prefix('@-slideshow-fullscreen')); // viewport unit fix for height:100vh - should be fixed in iOS 8
 
             this.slides.each(function(index) {
 
@@ -159,7 +160,14 @@
             this.slides.eq(this.current).addClass(UI.prefix('@-active'));
             this.triggers.filter(UI.prefix('[data-@-slideshow-item="'+this.current+'"]')).addClass(UI.prefix('@-active'));
 
-            UI.$win.on("resize load", UI.Utils.debounce(function() { $this.resize(); }, 100));
+            UI.$win.on("resize load", UI.Utils.debounce(function() {
+                $this.resize();
+
+                if ($this.fixFullscreen) {
+                    $this.container.css('height', window.innerHeight);
+                    $this.slides.css('height', window.innerHeight);
+                }
+            }, 100));
 
             this.resize();